sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ward-protocol"
version = "0.1.0"
description = "Official Python SDK for Ward Protocol - XRPL AMM DeFi"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Ward Protocol Team", email = "team@wardprotocol.com"}
]
keywords = ["xrpl", "defi", "amm", "crypto", "blockchain"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "xrpl-py>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[project.urls]
Homepage = "https://github.com/ward-protocol"
Repository = "https://github.com/ward-protocol/ward-protocol"

[tool.setuptools.packages.find]
where = ["."]
include = ["ward_protocol*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
